requirements.txt
arrow==0.10.0
kafka-python==1.3.5
